\name{calc.tree.sapwood.area}
\alias{calc.tree.sapwood.area}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calculate individual tree sapwood area in cm2}
\usage{
calc.tree.sapwood.area(DBH, sapwood.depth)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DBH}{The diamater at breast hieght (1.37m) of the tree. DBH units: cm. }
  \item{sapwood.depth}{The maximum depth of tree sapwood in cm. }
}
}
\value{
  \item{sapwood.area}{Tree sapwood area in cm2}
}
\author{
  Heather Speckman (HeatherNSpeckman@gmail.com)
}
