\name{convert.sapflux.gperm2ground.to.mm}
\alias{convert.sapflux.gperm2ground.to.mm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Convert sap flux units to mm }
\description{The function converts sap flux from units of ( g / m2 ground area / time ) to ( mm / time)}
\usage{
convert.sapflux.gperm2ground.to.mm(sapflux.data.m2ground)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sapflux.data.m2ground}{ Sap flux data in per m2ground.  Example: ( g / m2ground / s)}
}
\author{
  Heather Speckman (HeatherNSpeckman@gmail.com)
}
