\name{convert.sapflux.m2sapwood.to.m2ground}
\alias{convert.sapflux.m2sapwood.to.m2ground}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert sapflux area units
}
\description{
Convert sapflux area units from (m2 of sapwood) to (m2 of ground).
}
\usage{
convert.sapflux.m2sapwood.to.m2ground( sapflux.data,
                                       sapwood.per.ground.area )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sapflux.data}{ A data.frame or matrix of sap flux data.}
  \item{sapwood.per.ground.area.data}{Sapwood per ground area.  Must be in units (cm2 sapwood) / (m2 ground)}
}
\author{
  Heather Speckman (HeatherNSpeckman@gmail.com)
}
