\name{convert.sapflux.time.units}
\alias{convert.sapflux.time.units}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{convert.sapflux.time.units
}
\description{
Convert sapflux time units
}
\usage{
convert.sapflux.time.units(sapflux.data,
                           starting.time.units,
                           disired.time.units)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sapflux.data}{ A data.frame or matrix of sap flux data}
  \item{starting.time.units}{ Existing time units.
  "s" = second, "min" = minute, "hr" = hour, "day" = day}
  \item{disired.time.units}{ Disired time units}
}
\author{
Heather Speckman (HeatherNSpeckman@gmail.com)
}
